+++ iocore/eventsystem/UnixEThread.cc
@@ -41,7 +41,7 @@ struct AIOCallback;
 #define NO_ETHREAD_ID                   -1
 
 EThread::EThread()
- : generator(time(NULL) ^ (long) this),
+  : generator((uint32)((uintptr_t)time(NULL) ^ (uintptr_t) this)),
    ethreads_to_be_signalled(NULL),
    n_ethreads_to_be_signalled(0),
    main_accept_index(-1),
@@ -54,7 +54,7 @@ EThread::EThread()
 }
 
 EThread::EThread(ThreadType att, int anid)
-  : generator(time(NULL) ^ (long) this),
+  : generator((uint32)((uintptr_t)time(NULL) ^ (uintptr_t) this)),
     ethreads_to_be_signalled(NULL),
     n_ethreads_to_be_signalled(0),
     main_accept_index(-1),
@@ -90,7 +90,7 @@ EThread::EThread(ThreadType att, int anid)
 }
 
 EThread::EThread(ThreadType att, Event * e, ink_sem * sem)
- : generator(time(NULL) ^ (long) this),
+ : generator((uint32)((uintptr_t)time(NULL) ^ (uintptr_t) this)),
    ethreads_to_be_signalled(NULL),
    n_ethreads_to_be_signalled(0),
    main_accept_index(-1),
